Move transformation into user job.
